fix: macro for enemy spawn

+++ server/src/Server.hpp
@@ -2,6 +2,7 @@
 
 #include <asio.hpp>
 #include <chrono>
+#include <cstddef>
 #include <cstdint>
 #include <memory>
 #include <vector>
@@ -76,7 +77,7 @@ namespace server {
       void processGameEvents();
       void handleGameEvent(const queue::GameEvent &event);
 
-      int findExistingClient();
+      size_t findExistingClient();
       void handlePlayerInfoPacket(const char *data, std::size_t size);
       void handleClientData(std::size_t client_idx, const char *data,
                             std::size_t size);
